rockchip: vendor: Fix 'ulong' to 'int'
ulong offset is always >= 0 and it makes:
"WARNING: could not set reg FDT_ERR_BADOFFSET."

+++ arch/arm/mach-rockchip/vendor.c
@@ -578,7 +578,7 @@ void vendor_storage_fixup(void *blob)
 {
 	unsigned long size;
 	unsigned long start;
-	ulong offset;
+	int offset;
 
 	/* init vendor storage */
 	if (!bootdev_type) {
